requirements.txt
tensorflow_addons==0.11.2
tensorflow==2.3.1
Flask==1.1.2
boto3==1.16.6
numpy==1.18.5
pandas==1.1.3
mysql_connector_repackaged==0.3.1
scikit_learn==0.23.2
